Using 'gem contents' to figure out the directory where the gem is installed wasn't working with the ruby-oci8 gem (paths returned were relative to that directory), so switch to an alternate approach. This new method isn't particularly any more elegant, so it too may break at some point.
Only call refresh_metadata when needed (which is only in one place), not
in Tpkg::init. Loading the native package data takes many seconds on
platforms with lots of native packages, which introduces unnecessary and
unwelcome delay if that metadata is not needed.

Bug fix for packages not at base URL

Explicitely set permissions on tpkg.conf, and cron.d/tpkg on Red Hat.
We were setting the permissions on other files in the package, just not 
those two.  Cron is particular about the permissions of crontab files,
and will ignore them if the permissions aren't acceptable.