Update copyright message to include 2010.
Remove a check for :type not being set in a req, as a previous commit
ensured that :type is set in all reqs that are generated.

Catch and re-raise exceptions throw while running externals scripts,
adding the name of the external and associated package file to make it
easier for users to debug problems associated with externals.

Fix up the logic in check_requests.  It was returning the correct
result, but often not displaying the correct messages when a request
could not be satisfied.

Fix a case there :type in a req was checked for a :native_installed
value, which is not a valid value for :type.