Change default gem command from /home/t/ruby/bin/gem to /home/t/ruby-1.8/bin/gem to match our recommendation for how folks package major infrastructure apps.
Replace a few other constants related to paths with code to query the
gem command for those paths.  Eliminates the mess of confusing constants
at the top of the executable and allows users to actually package using
a different gem/ruby without a lot of hassle.

Figure out the name of the package that contains the gem command from
tpkg, rather than relying on a constant.

Allow the user to specify a string to add to the package name via a new
--extra-name option.  This is particularly useful if the user is
packaging gems against a non-standard copy of Ruby, they can customize
the package name to differentiate it from a package of that same gem
packaged for the site's default ruby.

If the user doesn't specify --extra-name, but does specify a non-default
gem command then insert a variation of the package name that contains
that non-standard gem into the package name.  Same reason as
--extra-name, differentiate these packages from packages using the
default gem.

Remove the code which renamed packages to incorporate the OS and
architecture, now that tpkg does that itself.

Removed some old commented out code.