This change is to stop setting GEM_PATH before installing the requested gems.  My logic for this change (written before making the change, thus the odd tense) is below:
gem2tpkg currently sets GEM_PATH to just the gem2tpkg working directory.
This ensures that gem does not find any gems installed in the standard
location. The advantage of this is that the user is able to package any
gem, even if they already have it installed. If GEM_PATH were not set
then gem would possibly see that a gem is already installed and skip it.

However, setting GEM_PATH turns out to cause a problem, at least with
building the xmlparser gem. That gem uses rake as part of its build
process. If GEM_PATH is set then rake is unable to find its own
libraries and fails.

I'm waffling on the best way to fix this. I think I'm probably
overestimating the need/benefit of setting GEM_PATH and that the default
ought to be to not set it. Particularly since I think gem's default
behavior when told to install a gem is to re-install it even if the gem
is already installed. If that proves troublesome in some circumstances
we can add an option to allow the user to request that GEM_PATH be set.