When we catch and re-raise exceptions (in order to add additional info to the exception message) from user code (externals, etc.) set the re-raised exception's backtrace to the original backtrace so that the backtrace info is not lost.  Added method self.wrap_exception to handle this.
Added a cache to available_packages_that_meet_requirement.  My primary
reason is to simplify the debug output associated with dependency
resolution (since it cuts out a bunch of duplicate runs through the
algorithm to calcuate the answer), but it should also speed things up
somewhat.

Check the exit value of externals and raise an exception if they fail.

Rearrange run_externals_for_install so that the check against
externals_to_skip is done earlier.  There's no point in doing any
processing of an external if we're going to skip it.

Check the exit value of externals datascripts and raise an exception if
they fail.

Add a begin/ensure block to run_externals_for_install to ensure that pwd
is restored in the event of errors.