In class Tpkg hange CONFIGDIR to DEFAULT_CONFIGDIR, and compute @configdir in initialize method using a similar algorithm as for @base. This allows the test suite to use a private config directory and not be affected by any files in the system config directory.
Change gethttp method from class to instance method so that it can 
utilize @configdir.

Make the :base option to initialize actually optional, setting it to 
DEFAULT_BASE if it is not specified.  The tpkg executable is essentially 
doing that as well, but other users of the library would benefit from 
the same logic.