Switch from Tempdir to Dir.mktmpdir.
Update test_init_links and test_crontab_destinations to handle the fact
that tpkg now just warns if run on a platform for which tpkg has no init
script or crontab support.  It used to raise an exception.  I had in
mind some fancier fix here in the test suite and had just been ignorning
the test failures.  But no practical fancy solution has come to me and
I'm tired of having to remember that these were "expected" failures.