Initialize @file_path to nil in the initialize method
Fix to_hash to not assume XML as the format if the format is not YAML.
We might add support for other formats in the future. As with other
methods in this class it now raises an exception if the format is an
unexpected one.

Heavily modify the add_tpkg_version method.
- Rearrange the method overall to DRY things up.
- XML metadata no longer directly parsed via REXML in order to read it.
  Use our [] functionality instead to DRY that section up.
- Use the version parameter rather than Tpkg::VERSION. This class
  shouldn't depend on Tpkg. The only call to this method already passed
  Tpkg::VERSION in the version parameter, but the version parameter was
  ignored by this method in favor of Tpkg::VERSION. So no behavior changes
  at this time.
- Add tpkg_version to in-memory data. Previously this method only
  updated the metadata file.
- Properly handle the case where this method is called and we don't know
  the file associated with this metadata.
- API change: no longer rescues Errno::EACCES. Handling that should be
  up to the caller.

Rearrange/clean up get_native_deps. API change: it now always returns an
array. Previously it would sometime (but not always) return nil instead
of an empty array if there were no native dependencies.