Add a "sudo" setting to tpkg.conf.
Set sudo usage based on the config file setting if not set by the user
on the command line.

Use the sudo_default? method from tpkg.rb rather than local logic to
decide the default value for sudo usage.  The sudo_default? method uses
the same logic as what we were previouly using here, so there is no
change in behavior on that front.